models/sbv2_core/requirements.txt
# sbv2_core base requirements (torch 제외)
# - RTX 50xx 등 환경 차이로 torch는 별도 설치를 권장.

accelerate
cmudict
cn2an
g2p_en
GPUtil
gradio>=4.32
jieba
librosa==0.9.2
loguru
nltk<=3.8.1
num2words
numpy<2
onnx
onnxconverter-common
onnxsim-prebuilt
onnxruntime
onnxruntime-directml; sys_platform == 'win32'
onnxruntime-gpu; sys_platform != 'darwin'
protobuf==4.25
psutil
punctuators
pydantic>=2
pyloudnorm
pyopenjtalk
pyopenjtalk-dict
pypinyin
pyworld-prebuilt
jaconv
scipy
soundfile
stable_ts
tensorboard
transformers==4.36.2
umap-learn

# build/runtime helpers for pyopenjtalk portability
setuptools<81
wheel>=0.43
cmake>=3.20

# training stack
pyannote.audio==3.1.1
torchmetrics==0.11.4
pytorch-lightning==2.1.3
lightning-fabric==2.1.3
